Use bunder-generated bundle binstub
The bin/bundle binstub Rails generates is not compatible with newer
versions of Bundler (see issue: rails/rails#31193) so if you generate
binstubs e.g. for rspec after installing and try to use them, you get
this error:

"Your `bin/bundle` was not generated by Bundler, so this binstub cannot
run."

This is fixed on Rails master (see PR: rails/rails#33202) but the change
has not yet been released.

In the meantime, regenerating the binstub with Bundler fixes this:

`bundle binstub bundler --force`