Make db isolation spec more strict
The previous version allowed for a PR to contain database changes and
changes outside of the app/ directory that could influence the
application (e.g. config/ changes).

This change makes it so the spec will only allow changes to db/migrate
files if all of the other changes are inside db/ and spec/.