Extract page creation procedure
The slug needs to always be generated from the title and having that
logic exist only in the controller seems wrong.

One option is to add it as a before_save callback to the model but
experience has shown that active record callbacks are usually best
avoided.

This commit extracts this small procedure into a separate object.

I refer to this as a "procedure" instead of using the more common name
"service" object since I think it's slightly less vague: this object
encapsulates a specific procedure for creating a page.

This commit also adds a Result::Success and Result::Failure class for
wrapping the results of such procedures, which is slightly more
expressive than the more typical convention of returning true for
success/false for failures.

All of this is a little premature for what logic exists here. This
change is betting that this procedure will become more complex in the
future and that other procedures will emerge so it's useful to
establish a pattern.