Ignore file extensions in page slugs
By default controller actions implicitly respond to all formats so if
you make a request ending in ".js", Rails tries to be helpful and
processes it like the requested format is Javascript.

This results in the page being rendered but without the full layout.

This commit makes it so requests are always for the HTML format.